+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 Scalyr Agent 2 Changes By Release
 =================================
 
+## 2.0.3 "Capricious Cat" - Nov 30, 2014
+
+<!---
+Packaged by Steven Czerwinski <czerwin@scalyr.com> on Nov 30, 2014 12:30 -0700
+--->
+
+This release is under active development, so the release date and feature set is TBD.
+
+New features:
+
+* The ``run_monitor.py`` takes a new option ``-d`` to set the debug level.
+
 ## 2.0.2 "Bashful Bovine" - Oct 31, 2014
 
 <!---